+++ CHANGELOG.md
@@ -47,6 +47,10 @@ All generated files in `deps/` are encouraged to be version controlled.
 ### chore: Add the `nns-dapp` and `internet_identity` to the local canister IDs set by `dfx nns import`
 `dfx nns install` installs a set of canisters in a local replica.  `dfx nns import` complements this by setting the canister IDs so that they can be queried by the user.  But `dfx nns import` is incomplete.  Now it will also provide the IDs of the `nns-dapp` and `internet_identity` canisters.
 
+### feat: `.env` file includes all created canister IDs
+Previously the `.env` file only included canister IDs for canisters that were listed as explicit dependencies during the build process.
+Now all canisters that have a canister ID for the specified network are included in `.env`.
+
 ## Asset Canister Synchronization
 
 Added more detailed logging to `ic-asset`. Now, when running `dfx deploy -v` (or `-vv`), the following information will be printed:
@@ -65,10 +69,15 @@ In order to allow larger changes without exceeding the per-message instruction l
 
 The asset canister now properly removes the v2-certified response when `/index.html` is deleted.
 
-The HttpResponse type now explicitly mentions the `upgrade : bool` field instead of implicitly returning `false` all the time.
+The HttpResponse type now explicitly mentions the `upgrade : Option<bool>` field instead of implicitly returning `None` all the time.
 
-- Module hash: 44b969d29a35ab67da654b37895cebfdd8b390f4e8feef025b081a311245f36d
+- Module hash: 651425d92d3796ddae581191452e0e87484eeff4ff6352fe9a59c7e1f97a2310
+- https://github.com/dfinity/sdk/pull/3120
 - https://github.com/dfinity/sdk/pull/3112
+
+### Motoko
+
+Updated Motoko to 0.8.8
 
 # 0.14.0